+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Added support form for our new zammad ticket system !3210 !3873 @chriswalg @alex.simm
 - Add push notification toggle and prompt users to use push notifications for the chat #1967 !3375 @AntonBallmaier
 - Added store hygiene requirement #1508 !3756 @AntonBallmaier
+- Add data privacy notice to hygine quiz !3792 @AntonBallmaier
 
 ## Changes
 - Improve accessibility in create forum thread form !3488 @iron9